Move button and LED handling from UsbHidDevie to genreic Device class
